ENYO-1630: Fix selection issue in enyo/VirtualDataRepeater
Due to an oversight in a recent refactoring, selection was not
working properly in enyo/VirtualDataRepeater and its subkinds.

We were failing to update the childrenByIndex hash, which is used
in the overridden childForIndex() method, which in turn is used by
the selection logic inherited from enyo/DataRepeater.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)