ENYO-1461: Add `preventTap` option to configureHoldPulse
At the end of a hold, it is sometimes desirable to prevent the tap
event that would normally fire on `up`.

The Moonstone paging control was previously implementing its own
hold behavior, in part so that it could suppress the tap. The
other limitations that originally necessitated the custom hold
implementation had since been addressed, so adding `preventTap`
means that we can get rid of the custom implementation altogether.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)