change selected default to null
Using an out-of-range default value lets us know if a valid value is
set at create time. selected will be converted to a valid value on
render because it is updated from the DOM which will report the
selectedIndex of the selection <option> or 0 if none are selected.

Issue: ENYO-1553
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)