ENYO-1657: Another VDR selection fix
Follow-on to the fix for ENYO-1630.

We still had issues because we still weren't updating the
childrenByIndex hash in all circumstances.

Refactored so that we call `assignChild` regardless of whether
the model associated with a given control is changing.

Also: removed unused variable declaration.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)