ENYO-1664: Add `getFullyVisibleItems()` method
enyo/NewDataList has a `getVisibleItems()` method, which returns an
array of list children that are currently scrolled into view. This
array contains all elements that are at all visible, including
those that are only partially visible.

For some operations, consuming code may require a list of only
those elements that are fully visible. (For example,
moonstone/Scrollable has a need in certain circumstances to
redirect Spotlight focus from a hidden or partially visible child
to a fully visible child, to avoid unwanted auto-scrolling.) We
therefore add a separate method for this purpose.

Like the rest of NewDataList, these APIs are to be considered
experimental and subject to change. Depending on observed use
cases, we may ultimately decide to make this functionality more
configurable and/or expose it via one method vs. several.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)