ENYO-1689: Accessibility: migration accessibilityDisabled to enyo 2.6
The accessibilityDisabled is to be disable voice readout for control.
In some requirements, it should not read anything when control is
focused, so we provide accessibilityDisabled property to cover that.
If accessibilityDisabled is true, aria attributes are removed for the
control.
Note that this is not working on HTML form elements which can get focus
without tabindex.

https://jira2.lgsvl.com/browse/ENYO-1689
Enyo-DCO-1.1-Signed-off-by: Jaewon Jang <jaewon98.jang@lgepartner.com>