ENYO-2083: Documenting enyo/History as WIP
The refactoring of moon/History (including creation of the new
enyo/History singleton) fixes multiple known issues and is a
significant improvement, but I'm not confident that we have things
quite right yet. I therefore want to mark enyo/History as an
experimental / work-in-progress API for the time being to allow us
to make breaking changes or replace the API altogether in
subsequent releases.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)