ENYO-2313: In enyo/Image, ensure that placeholder image fits
At least in cases where the image size was not explicitly
specified and the sizing option was not being used, the placeholder
image was not properly scaled to the size of the image element.

Some simple CSS refactoring takes care of that.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)