ENYO-2452-rajyavardhan-Unnecessary border around placeholder if image fails to load
Issue: The behavior is rather normal as enyo.Image makes the 'src:
null', if the image fails to load. So browser draws a border around the
size of the image when there is no 'src' attribute or 'src: null'
Fix: Instead of making the src: null when the image doesn't load, the
source is set with a transparent image.
Enyo-DCO-1.1-Signed-off-by: Rajyavardhan P <rajyavardhan.p@lge.com>