ENYO-2258: Defer reflow when hidden, do it when shown
Reflowing a Control when it's hidden generally doesn't make sense,
since a) it's wasteful to spend cycles laying out something that
can't be seen; and b) the layout may depend on measuring the
Control's children, and they can't be accurately measured if
they're not showing.

We therefore defer reflow in the case where we're hidden, and
wait until we're shown.

In the case where we are explicitly hidden (i.e., `showing` is
`false`), we defer immediately. We can't inexpensively detect the
case where we're showing but an ancestor is hidden, so to handle
that case we rely on the layout to do its measurements and let
us know (by returning `true`) if we need to reflow again the next
time we are shown.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)