ENYO-2565: Bullet-proof logic for constraining min threshold
In `NewDataList`, the logic for constraining the minimum scroll
threshold value was susceptible to failure in the case where the
values being compared were long floating point values (we were
checking for equality, but in the failure case two values that were
"conceptually equal" for the purpose of our algorithm were actually
different in digits far to the right of the decimal). We replace
the equality operator  with a call to `Math.min()` to constrain the
threshold  value, which gives us the desired result.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)