ENYO-2644: Override scrollToControl() in enyo/NewDataList
enyo/Scrollable provides a `scrollToControl()` method that
has a generalized implementation for calculating a Control's
position and scrolling to it.

In the case of enyo/NewDataList, it's possible to provide an
alternative implementation that is more performant and more
accurate (more performant because it doesn't need to query the DOM
for bounds, more accurate because the generalized implementation
is influenced by current scroll position and may give slightly
different values when animating vs. not).

We tweak enyo/Scrollable to defer to an alternative method when one
is provided, and we provide an alternative method in
enyo/NewDataList.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)