reset the start position on render for translateOptimized scrollers
When re-rendering a scrolled scroller using TranslateScrollStrategy
with translateOptimized: true, the startX/Y values would be updated to
be a non-zero value matching the prior values of scrollLeft/Top.
However, the scrollNode was also translated the same amounts. As a
result, when the user scrolled, the content would jump the amount of
startX/Y resulting in whitespace above the content.

Since the purpose of startX/Y is to offset the transform the amount of
the scrollLeft/Top of the scrollNode but those values are never
updated with translateOptimized: true, we'll reset them to 0 in
rendered() so no offsetting occurs.

Issue: ENYO-2691
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)