simply declarative child ViewManager handling
Rather than checking for the isManager flag which was inconsistently
applied, a VM can call a private isManager method to determine if the
provided view exists and is in its viewManagers hash. A consequence of
this change is that VMs can only be declared within the managers[] and,
when created manually, will always be considered a manager.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)