ENYO-2644: Scrollable: Snap to boundaries
Often, if you're scrolling to a target near the edge of a
scrollable region, it makes sense to scroll all the way to
the edge.

To support this use case, we add an (experimental, for now)
option called `boundarySnapThreshold` to ScrollMath.

At the same time, we do a bit of minor code cleanup on another
experimental feature of a similar nature (which supports snapping
to the boundaries of items within the scrollable region in the
specific case where the items are evenly sized and spaced). Due
to time constraints and the fact that these features are not
ready for wide use, leaving docs as a TODO for the moment.

This change was precipitated by an issue filed against Moonstone,
where 5-way scrolling to the first / last item in a list was not
scrolling all the way to the scroller boundary; as a result, the
up / down arrows in the scroll controls were not being disabled,
making it unclear that the end of the list had been reached.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)