Revert "ENYO-2644: Override scrollToControl() in enyo/NewDataList"
This reverts commit 37b81a69844807ba5ed16f81b7797ae16dadf8a9.

The original fix for ENYO-2644, which we're reverting here, didn't
really address the issue. It made the symptoms disappear, but
not in the way originally intended because the issue wasn't
properly diagnosed. It also introduced some new issues because
it unintentionally skipped some logic related to scrolling options.

There were actually some performance- and customization-related
benefits to the change, so we'll likely want to reintroduce a
version of it at some point in the future, but we'll simply revert
it for now since we are at a point in the release cycle where we
need to minimize risk of regressions.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)