ENYO-2696: Restore async to VirtualDataRepeater.refresh()
In the original fix for ENYO-2514, we modified VDR.refresh() to
make it work synchronously, as the historical reasons for making it
async (inherited from DataRepeater) seemed not to apply, and the
async behavior was causing a problem in one case: when the refresh
was deferred because the repeater was hidden, the "stale" state of
the repeater would flash momentarily on screen after the repeater
was re-shown but before the refresh occurred.

Making refresh() synchronous fixed that case, but it appears that
there are still some cases where async refresh has significant perf
benefits, so in this change we're restoring the async behavior and
taking a different approach to fixing ENYO-2514.

The new issue (ENYO-2696) that prompted the change back to async
was one in which the collection underlying the repeater was emptied
and immediately repopulated. In this case, the async approach
results in debouncing the refresh; it only refreshes once, after
repopulating, rather than once after emptying and again after
repopulating.

To solve the flashing-stale-content issue, we now use the CSS
visibility property to make the repeater invisible while it's in
the process of performing a deferred refresh.

While making this change, also did some cleanup and added some
inline comments.

Finally, note that the logic for deferring refresh() and reset()
originally lived in NewDataList, not in VirtualDataRepeater. In
the process of reviewing and integrating the original fix for
ENYO-2514, we determined that it made sense to move it into VDR
and wrote up a new issue (ENYO-2533) to track that work. Since
we're back in the same code now, we're making that change, too.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)