support multiple calls to activate() in the same frame
View activation is asynchronous (using rAF) so multiple calls to
activate() in the same frame produced unexpected results and out of
sync styling. For floating panels, the stack would only include the
last view. For fixed panels, the direction could be incorrect if the
penultimate activation was on the opposite side of the final activated
view from the source view.

Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)