ENYO-2744: Tweaks to allow restoration of list state on render
We have recently started getting more aggressive about view
virtualization for performance reasons. One tactic in this area
involves tearing down the DOM representation of a Control and then
restoring it later as needed ("view caching").

In the case where the view being cached / restored contains a
DataList, the state of the view was not being preserved as expected
because DataList effectively resets upon being (re-)rendered.

The proper fix for this issue would involve rethinking the DataList
lifecycle with view-virtualization scenarios in mind, but this
would be a significant effort and the current DataList is on a
relatively short road to replacement, so for now we'll go with a
lighter-weight, Moonstone-specific solution.

That solution does, however, require a number of minor tweaks and
enhancements in Enyo:

* A minor change to the delegation interface between enyo/Control
  and enyo/HTMLStringDelegate, allowing `teardownChildren()` to be
  more easily extended or overridden.

* A new `callback` argument for `enyo/DataList.scrollToIndex()`,
  allowing for actions to be performed after this potentially
  asynchronous operation is completed.

* A change in `enyo/ViewPreloadSupport.cacheView()` to defer
  removing the view's DOM node until after its children have been
  torn down. This enables views to hook `teardownChildren()` and
  do things in preparation for being torn down that take current
  DOM state into account.

* To support the above change in enyo/ViewPreloadSupport, a minor
  refactoring of some cross-platform node-removal logic, moving it
  from enyo/Control to enyo/dom.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)