ENYO-2788: Provide an explicit `beforeTeardown()` hook for Control
In the fix for ENYO-2744, we made a tweak to the interface between
enyo/Control and enyo/HTMLStringDelegate to make it possible for
Controls to override `teardownChildren()`. In doing so, we failed
to preserve the `cache` argument to `teardownChildren()`, which
caused a regression for Controls that use the 'renderOnShow'
feature.

This regression could have been addressed by simply making sure we
passed the `cache` argument properly, but on further reflection it
seems safer and more future-proof to instead revert the previous
change in favor of an explicit `beforeTeardown()` lifecycle method
that Controls can implement as needed. Doing that now.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)