ENYO-2783: Refactoring in enyo/Scrollable to allow customization
The logic for scrolling to a child of Scrollable has until now
been essentially monolithic, but we now refactor it to allow parts
of the logic to be selectively overridden.

Specifically, there are now separate methods for getting the
offset coordinates of the child (`getChildOffsets()`), calculating
the target scroll coordinates (`getTargetCoordinates()`), and
actually effecting the scroll (`scrollToTarget()`).

The immediate use case is to provide a custom implementation of
`getChildOffsets()` for NewDataList, since NewDataList can get
child offsets more efficiently than the default implemenation does,
and can also properly account for the spacing of list items in its
own implementation. (Not properly accounting for spacing was the
actual root cause of ENYO-2644 and ENYO-2783, the issues that led
to this work. As part of this pull request, we are also reverting
earlier, flawed solutions to the problem described in ENYO-2644).

Note that we also have a use case for overriding `scrollToTarget()`
(see ENYO-2710 and ENYO-2729), but that work is not required to fix
ENYO-2783 so we aren't doing it now.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)