ENYO-2840: Make `accessibilityPreventScroll` work in RTL
The `accessibilityPreventScroll` feature prevents the browser
engine from doing bad auto-scrolling things when an item is focused
in one of our scrollers. It works by listening for native scroll
events and forcibly resetting native scrollTop and scrollLeft
properties as needed.

Our logic for resetting scrollLeft did not previously account for
RTL, so we fix that.

As noted inline, the current fix is specific to Blink and WebKit;
we'll need to do something clever to make this work cross-browser,
but not attempting that now because cross-browser support is not
urgent for the imminent internal release, and this is not our only
cross-browser scrolling issue. Wrote up ENYO-2841 to track future
work.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)