Revert "ENYO-2808 Clear metrics to allow for page generation"
This reverts commit 6d477c1c5403b131f229d7888fe397f862e5225f.

Reverting this change because it turns out not to be necessary to
address any currently known issue, and we're at a point in both
the product release cycle and the lifespan of enyo/DataList where
minimizing risk of new regressions is our top priority. Leaving in
the git log, though, in case we run into problems in this area and
want to retrace our steps.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)