guard against shifting onhold events for cancelled holds
If endHold() is called during a hold event handlers, this._unsent will
be nulled which causes an exception because it is assumed to be an
array. Adding a truthy guard to ensure that the next event is only
shifted if it hasn't been cleared.

Issue: ENYO-3038
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)