guard getRelations against nulled this.relations
If a RelationalModel particpates in a binding, it may be accessed after
it is destroyed. Because it overrides get() to interrogate its related
models, if it had been destroyed the array of models would have been
nulled causing a runtime error.

Issue: ENYO-2990
Enyo-DCO-1.1-Signed-off-by: Ryan Duffy (ryan.duffy@lge.com)