ENYO-3219: Refresh scroll thresholds when adding / removing models
If a list is scrolled to the end and the size of the collection
changes, we need to recalculate the scroll thresholds and possibly
adjust the range of items to be rendered.

Until now, we've only adjusted thresholds in the `scroll()` method,
based on the direction and distance we've scrolled. In this fix, we
add a new method, `refreshThresholds()`, which recalculates the
thresholds based on the current scroll value. We call the new
method from `modelsAdded()` and `modelsRemoved()` to make sure that
the thresholds and item range are correct after the size of the
collection has changed.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton (gray.norton@lge.com)