ENYO-3382: Add support showing and hiding method to transition support
When we apply showingTransitionSupport on control, it is hard to hook
before showing start timing from app side. This is because the showing
changed handler is calling super call when animation ends.

We add showingMethod and hidingMethod api to support this case.

Enyo-DCO-1.1-Signed-off-by: Kunmyon Choi (kunmyon.choi@lge.com)